+++ client/components/TestPlanReportStatusDialog/index.jsx
@@ -7,7 +7,7 @@ import AddTestToQueueWithConfirmation from '../AddTestToQueueWithConfirmation';
 import { useQuery } from '@apollo/client';
 import { ME_QUERY } from '../App/queries';
 import { evaluateAuth } from '../../utils/evaluateAuth';
-import getMetrics from './getMetrics';
+import getMetrics from '../Reports/getMetrics';
 import { calculateTestPlanReportCompletionPercentage } from './calculateTestPlanReportCompletionPercentage';
 
 const TestPlanReportStatusModal = styled(Modal)`
